third_party/WebKit/LayoutTests/http/tests/locks/lock-attributes.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Web Locks API: Lock Attributes</title>
<link rel=help href="https://github.com/inexorabletash/web-locks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(async t => {
  await navigator.locks.acquire(['r1', 'r2'], lock => {
    assert_array_equals(lock.scope, ['r1', 'r2']);
    assert_equals(lock.mode, 'exclusive');
  });
}, 'Lock attributes reflect requested properties (exclusive)');

promise_test(async t => {
  await navigator.locks.acquire(['r1', 'r2'], {mode: 'shared'}, lock => {
    assert_array_equals(lock.scope, ['r1', 'r2']);
    assert_equals(lock.mode, 'shared');
  });
}, 'Lock attributes reflect requested properties (shared)');

promise_test(async t => {
  await navigator.locks.acquire(['r2', 'r1'], lock => {
    assert_array_equals(lock.scope, ['r1', 'r2']);
  });
}, 'Lock.scope is in sorted order');

promise_test(async t => {
  await navigator.locks.acquire(['r1', 'r2', 'r1', 'r2'], lock => {
    assert_array_equals(lock.scope, ['r1', 'r2']);
  });
}, 'Lock.scope has no duplicates');

</script>
